# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "backend server of movie ticketing NOIR Project",
        "title": "NOIR RESTful API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/transaction/": {
            "post": {
                "security": [
                    {
                        "Token": []
                    }
                ],
                "description": "Create a new movie ticket transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transaction"
                ],
                "summary": "Create a new transaction",
                "parameters": [
                    {
                        "description": "Transaction details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateTransactionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Transaction created successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.TransactionResult"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/transaction/payment": {
            "post": {
                "security": [
                    {
                        "Token": []
                    }
                ],
                "description": "Process a payment for a transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transaction"
                ],
                "summary": "Process payment",
                "parameters": [
                    {
                        "description": "Payment details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ProcessPaymentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment processed successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.TransactionResult"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/transaction/{code}": {
            "get": {
                "security": [
                    {
                        "Token": []
                    }
                ],
                "description": "Get full details of a transaction using its unique code",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transaction"
                ],
                "summary": "Get transaction by code",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Transaction Code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.TransactionListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Transaction not found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.CinemaResponse": {
            "type": "object",
            "properties": {
                "cinema_id": {
                    "type": "integer"
                },
                "location": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "dto.CreateTransactionRequest": {
            "type": "object",
            "required": [
                "payment_method_id",
                "recipient_email",
                "recipient_full_name",
                "recipient_phone_number",
                "seat_numbers",
                "showtime_id"
            ],
            "properties": {
                "payment_method_id": {
                    "type": "integer"
                },
                "recipient_email": {
                    "type": "string"
                },
                "recipient_full_name": {
                    "type": "string"
                },
                "recipient_phone_number": {
                    "type": "string"
                },
                "seat_numbers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "showtime_id": {
                    "type": "integer"
                }
            }
        },
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {},
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "dto.MovieResponse": {
            "type": "object",
            "properties": {
                "backdrop_path": {
                    "type": "string"
                },
                "cast": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "director": {
                    "type": "string"
                },
                "duration": {
                    "type": "integer"
                },
                "genre": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "movie_id": {
                    "type": "integer"
                },
                "overview": {
                    "type": "string"
                },
                "poster_path": {
                    "type": "string"
                },
                "release_date": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "dto.ProcessPaymentRequest": {
            "type": "object",
            "required": [
                "transaction_code"
            ],
            "properties": {
                "payment_proof": {
                    "type": "string"
                },
                "transaction_code": {
                    "type": "string"
                }
            }
        },
...
        "dto.ShowtimeResponse": {
            "type": "object",
            "properties": {
                "base_price": {
                    "type": "number"
                },
                "show_datetime": {
                    "type": "string"
                },
                "showtime_id": {
                    "type": "integer"
                }
            }
        },
        "dto.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "dto.TicketResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "seat_number": {
                    "type": "string"
                },
                "showtime_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "ticket_code": {
                    "type": "string"
                },
                "ticket_id": {
                    "type": "integer"
                },
                "transaction_id": {
                    "type": "integer"
                }
            }
        },
        "dto.TransactionListResponse": {
            "type": "object",
            "properties": {
                "cinema": {
                    "$ref": "#/definitions/dto.CinemaResponse"
                },
                "created_at": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "movie": {
                    "$ref": "#/definitions/dto.MovieResponse"
                },
                "seats": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "showtime": {
                    "$ref": "#/definitions/dto.ShowtimeResponse"
                },
                "status": {
                    "type": "string"
                },
                "total_amount": {
                    "type": "number"
                },
                "transaction_code": {
                    "type": "string"
                },
                "transaction_id": {
                    "type": "integer"
                }
            }
        },
        "dto.TransactionResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "integer"
                },
                "expires_at": {
                    "type": "string"
                },
                "paid_at": {
                    "type": "string"
                },
                "payment_method_id": {
                    "type": "integer"
                },
                "recipient_email": {
                    "type": "string"
                },
                "recipient_full_name": {
                    "type": "string"
                },
                "recipient_phone_number": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "total_amount": {
                    "type": "number"
                },
                "total_seats": {
                    "type": "integer"
                },
                "transaction_code": {
                    "type": "string"
                },
                "transaction_id": {
                    "type": "integer"
                }
            }
        },
        "dto.TransactionResult": {
            "type": "object",
            "properties": {
                "tickets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.TicketResponse"
                    }
                },
                "transaction": {
                    "$ref": "#/definitions/dto.TransactionResponse"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Token": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}